Add a 'github_banner_image' option.
The 'github_banner_image' option allows the user to specify an image to
be used as a replacement for the default 'Fork me on Github' banner. The
image path must be relative to `$PROJECT/_static/`. If this option is
unset (which is the default), then the standard 'Fork me on Github'
banner is used.