Improved admonition differentiation
Addresses #28 and #6.

To do this, explicit styling was applied for the different admonition
types.

The Sphinx RTD theme uses the same styling for the following groups:
- blue: note, custom admonitions, seealso
- red: danger, error
- yellow: attention, caution, danger
- green: tip, hint, important

Because Alabaster's default aesthetics are not as shiny, we use the
following instead:
- gray: note, custom admonitions, seealso
- pink with darker pink shadow: danger, error
- pink: attention, caution, danger
- gray: tip, hint, important

In addition, the different admonition styles are separated into a better
set of comments for their readability.