Only load Google Analytics if user have not opt-in to DoNotTrack.
GA does not respect the "DoNotTrack" setting, so conditionally load it
only if it is not set by user.

Alternatively it appears that GA.js is deprected and has been replace by
analytics.js (according to official google Documentation [0]:

> ga.js is a legacy library. If you are starting a new implementation,
> we recommend you use the latest version of this library,
> analytics.js[1]. For existing implementations, learn how to migrate from
> ga.js to analytics.js[2].

But that will be the subject of another pull-request if maintainers
agrees that updating the theme from GA to analytics is desired.

0: https://developers.google.com/analytics/devguides/collection/gajs/
1: https://developers.google.com/analytics/devguides/collection/analyticsjs/
2: https://developers.google.com/analytics/devguides/collection/upgrade/