Use searchfield instead of searchbox component in sidebar
This moves the search field up before the navigation. It's quite
commmon that search is near the top of the page. Also, this
results in a fixed position, because navigation is page dependent
and may be long, which could move the search field far down or
even off-page in the old layout.

The searchfield component
(https://github.com/sphinx-doc/sphinx/pull/11045/)
is more compact and does not have a heading, which reduces
clutter and gives more focus to the content sections.